fix(export): distinguish empty isolation filter from no filter in OBJ export
The OBJ twin of the GLB fix in #4364 (the #4328 scenario): `ObjOptions::isolated`
(Rust) and `IfcAPI::export_obj`'s `isolated` parameter (wasm boundary) collapsed
"no isolation filter" and "isolation active, zero matches" into the same empty
value via `Vec::is_empty()`, so an active-but-empty filter silently exported the
whole model. `isolated` is now `Option<Vec<u32>>` on the Rust side and
`Uint32Array | undefined` on the TS side, matching the convention #4364
established for GLB. The CLI (`export-rust-formats.ts`) and MCP `export_obj`
tool both now pass `undefined` (not an empty Uint32Array) when no filter is
requested, avoiding the same regression #4364 had to fix for GLB in the same PR.

`obj.rs`'s inline tests moved to `obj_tests.rs` (`#[path]`, mirrors
`gltf_tests.rs`) to stay under the module-size ratchet after the new repro
tests.
