+++ packages/wasm/pkg/ifc-lite.d.ts
@@ -498,14 +498,21 @@ export class IfcAPI {
      * V8 max-string ceiling (~512 MB); decode with `TextDecoder` when a string
      * is genuinely needed.
      *
-     * `hidden` / `isolated` are express-id filters mirroring the viewer's visibility
-     * state (empty `isolated` ⇒ all visible). Instanced type-library shapes are skipped.
+     * `hidden` is an express-id filter mirroring the viewer's visibility state.
+     * `isolated` carries the isolation allowlist's null-vs-empty distinction across
+     * the wasm boundary: omit it (`undefined`) for "no isolation filter" (every mesh
+     * is a candidate); pass an empty `Uint32Array` for "isolation is ACTIVE and
+     * currently matches nothing" (every mesh is excluded). Collapsing the two — as a
+     * bare `Uint32Array` parameter would force a caller to do — silently exports the
+     * whole model when a filter matches nothing (the OBJ twin of #4328/#4364, fixed
+     * for GLB in `exportGlb`). A non-empty `Uint32Array` is the ordinary allowlist.
+     * Instanced type-library shapes are skipped regardless of the filter.
      *
      * ```javascript
-     * const obj = api.exportObj(ifcContent, true, new Uint32Array(), new Uint32Array());
+     * const obj = api.exportObj(ifcContent, true, new Uint32Array(), undefined);
      * ```
      */
-    exportObj(content: Uint8Array, include_normals: boolean, hidden: Uint32Array, isolated: Uint32Array): Uint8Array;
+    exportObj(content: Uint8Array, include_normals: boolean, hidden: Uint32Array, isolated: Uint32Array | null | undefined): Uint8Array;
     /**
      * Re-serialize the model in `content` to STEP/IFC UTF-8 bytes.
      *
