+++ rust/export/src/obj_tests.rs
@@ -1,0 +1,283 @@
+// SPDX-License-Identifier: MPL-2.0
+//! Tests for `obj.rs`, split out under the house pattern (AGENTS.md).
+//!
+//! `obj.rs` sits on the module-size ratchet's allowlist at its recorded
+//! budget, and the ratchet counts non-test lines -- so tests that grow with
+//! the module have to live beside it rather than inside it (mirrors
+//! `gltf_tests.rs` for the sibling GLB exporter).
+
+use super::*;
+
+#[test]
+fn duplex_exports_well_formed_obj() {
+    let (obj, stats) =
+        export_obj_with_stats(&fixture_or_skip!("ara3d/duplex.ifc"), &ObjOptions::default());
+    assert!(stats.meshes > 0, "expected meshes");
+    assert!(stats.vertices > 0, "expected vertices");
+    assert!(stats.triangles > 0, "expected triangles");
+    assert!(obj.contains("\nv "), "has vertices");
+    assert!(obj.contains("\nf "), "has faces");
+    assert!(obj.contains("\no Ifc"), "has element object groups");
+
+    // Every face index must reference a written vertex (1..=vertices).
+    let max_idx = stats.vertices;
+    for line in obj.lines().filter(|l| l.starts_with("f ")) {
+        for tok in line[2..].split_whitespace() {
+            let v: usize = tok.split("//").next().unwrap().parse().unwrap();
+            assert!(v >= 1 && v <= max_idx, "face index {v} out of range 1..={max_idx}");
+        }
+    }
+}
+
+#[test]
+fn isolation_filter_limits_output() {
+    let all = export_obj_with_stats(&fixture_or_skip!("ara3d/duplex.ifc"), &ObjOptions::default()).1;
+    // Find one express id that was emitted by re-reading meshes through the pipeline.
+    let result = process_geometry(&fixture_or_skip!("ara3d/duplex.ifc")[..]);
+    let some_id = result
+        .meshes
+        .iter()
+        .find(|m| super::mesh_visible(m, &None, &[]))
+        .map(|m| m.express_id)
+        .expect("at least one visible mesh");
+
+    let isolated = export_obj_with_stats(
+        &fixture_or_skip!("ara3d/duplex.ifc"),
+        &ObjOptions { isolated: Some(vec![some_id]), ..ObjOptions::default() },
+    )
+    .1;
+    assert!(isolated.meshes >= 1);
+    assert!(isolated.meshes <= all.meshes);
+}
+
+/// The OBJ twin of the #4328/#4364 GLB fix
+/// (`gltf_tests::mesh_visible_empty_isolated_set_is_indistinguishable_from_no_filter_bug_repro`):
+/// an ACTIVE isolation filter that matches nothing must hide every mesh,
+/// not export the whole model. Before the fix `ObjOptions::isolated` was a
+/// bare `Vec<u32>`, so "no filter" and "filter active, zero matches" both
+/// arrived as an empty vec and collapsed to the same `mesh_visible` outcome.
+/// This test pins that collapse; it must go RED before the fix (`isolated`
+/// becomes `Option<Vec<u32>>`-driven) and GREEN after.
+#[test]
+fn mesh_visible_empty_isolated_set_is_indistinguishable_from_no_filter_bug_repro() {
+    let mesh = MeshData::new(
+        99,
+        "IfcWall".to_string(),
+        vec![0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0, 0.0],
+        vec![0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0],
+        vec![0, 1, 2],
+        [1.0, 0.0, 0.0, 1.0],
+    );
+
+    let no_filter: Option<Vec<u32>> = None;
+    let empty_active_filter: Option<Vec<u32>> = Some(vec![]);
+
+    assert!(mesh_visible(&mesh, &no_filter, &[]), "no isolation filter: mesh stays visible");
+    assert!(
+        !mesh_visible(&mesh, &empty_active_filter, &[]),
+        "an isolation filter active with zero matches must hide every mesh, not export \
+         everything (before the fix both read as `isolated.is_empty() == true`, which \
+         can't tell 'no filter' from 'filter matched nothing')"
+    );
+}
+
+/// #4328/#4364 follow-up, three states over a real fixture: `None` (no
+/// filter) keeps the full model, an active filter matching zero express ids
+/// exports NOTHING (not the whole model), and an active filter matching one
+/// id exports exactly that id's mesh(es) — the middle case is the one a
+/// careless fix breaks in either direction (over-hiding a normal allowlist,
+/// or falling back to "export everything" on an empty one).
+#[test]
+fn isolation_three_states_over_a_real_fixture() {
+    let bytes = fixture_or_skip!("ara3d/duplex.ifc");
+
+    let no_filter = export_obj_with_stats(&bytes, &ObjOptions::default()).1;
+    assert!(no_filter.meshes > 1, "fixture must have more than one visible mesh to discriminate");
+
+    let empty_active = export_obj_with_stats(
+        &bytes,
+        &ObjOptions { isolated: Some(vec![]), ..ObjOptions::default() },
+    )
+    .1;
+    assert_eq!(
+        empty_active.meshes, 0,
+        "an ACTIVE isolation filter matching zero express ids must export zero meshes, \
+         not silently fall back to the whole model"
+    );
+
+    let result = process_geometry(&bytes[..]);
+    let some_id = result
+        .meshes
+        .iter()
+        .find(|m| super::mesh_visible(m, &None, &[]))
+        .map(|m| m.express_id)
+        .expect("at least one visible mesh");
+    let one_match = export_obj_with_stats(
+        &bytes,
+        &ObjOptions { isolated: Some(vec![some_id]), ..ObjOptions::default() },
+    )
+    .1;
+    assert!(
+        one_match.meshes >= 1 && one_match.meshes < no_filter.meshes,
+        "a non-empty allowlist must still export exactly its matches, not everything and not nothing"
+    );
+}
+
+/// #4056: the proper frame rotation must retain source triangle order.
+#[test]
+fn obj_faces_preserve_the_source_mesh_winding_4056() {
+    let bytes = fixture_or_skip!("ara3d/duplex.ifc");
+    let result = process_geometry(&bytes);
+
+    // `mesh_visible` only requires a NON-EMPTY index buffer, so a visible
+    // mesh may carry one or two indices and emit no face at all (the export
+    // loop uses `chunks_exact(3)`). Such a mesh still writes its vertices,
+    // advancing `vert_base` — so the first FACE need not belong to the first
+    // visible MESH, and its indices need not start at zero. Walk the same
+    // sequence the exporter walks and accumulate the offset, instead of
+    // assuming both.
+    let mut vert_base = 0usize;
+    let mut first = None;
+    for m in result.meshes.iter().filter(|m| mesh_visible(m, &None, &[])) {
+        if m.indices.len() >= 3 {
+            first = Some((m, vert_base));
+            break;
+        }
+        vert_base += m.positions.len() / 3;
+    }
+    let (mesh, vert_base) = first.expect("a visible mesh with a complete triangle");
+    let tri = &mesh.indices[0..3];
+    // OBJ indices are 1-based and global.
+    let idx = |i: u32| vert_base + i as usize + 1;
+    let expected = format!("f {} {} {}", idx(tri[0]), idx(tri[1]), idx(tri[2]));
+
+    let obj = export_obj(&bytes, &ObjOptions { include_normals: false, ..ObjOptions::default() });
+    let first_face = obj.lines().find(|l| l.starts_with("f ")).expect("a face line");
+
+    // A triangle whose 2nd and 3rd source indices coincide would make the
+    // reversal unobservable; assert the fixture is not that degenerate case.
+    assert_ne!(tri[1], tri[2], "fixture triangle must distinguish b from c");
+    assert_eq!(first_face, expected, "OBJ must preserve source winding");
+}
+
+/// #4056: inspect real exported positions, normals and both OBJ face syntaxes.
+#[test]
+fn obj_triangle_keeps_outward_face_orientation_4056() {
+    let bytes = br#"ISO-10303-21;
+HEADER;
+FILE_DESCRIPTION(('Orientation witness'),'2;1');
+FILE_NAME('triangle.ifc','2026-09-07T00:00:00',('Test'),('Test'),'ifc-lite','ifc-lite','');
+FILE_SCHEMA(('IFC4'));
+ENDSEC;
+DATA;
+#1=IFCCARTESIANPOINT((0.,0.,0.));
+#2=IFCDIRECTION((0.,0.,1.));
+#3=IFCDIRECTION((1.,0.,0.));
+#4=IFCAXIS2PLACEMENT3D(#1,#2,#3);
+#5=IFCLOCALPLACEMENT($,#4);
+#6=IFCGEOMETRICREPRESENTATIONCONTEXT($,'Model',3,0.00001,#4,$);
+#7=IFCSIUNIT(*,.LENGTHUNIT.,$,.METRE.);
+#8=IFCUNITASSIGNMENT((#7));
+#9=IFCPROJECT('0000000000000000000001',$,'Project',$,$,$,$,(#6),#8);
+#12=IFCCARTESIANPOINTLIST3D(((0.,0.,0.),(1.,0.,0.),(0.,1.,0.)));
+#13=IFCTRIANGULATEDFACESET(#12,((0.,0.,1.),(0.,0.,1.),(0.,0.,1.)),.F.,((1,2,3)),$);
+#14=IFCSHAPEREPRESENTATION(#6,'Body','Tessellation',(#13));
+#15=IFCPRODUCTDEFINITIONSHAPE($,$,(#14));
+#1000=IFCWALL('0000000000000000001000',$,'Triangle',$,$,#5,#15,$,.NOTDEFINED.);
+ENDSEC;
+END-ISO-10303-21;
+"#;
+    for include_normals in [false, true] {
+        let (obj, stats) = export_obj_with_stats(bytes, &ObjOptions {
+            include_normals, ..ObjOptions::default()
+        });
+        assert_eq!(stats.triangles, 1);
+        let vectors = |prefix: &str| -> Vec<Vec<f64>> {
+            obj.lines().filter_map(|line| line.strip_prefix(prefix))
+                .map(|line| line.split_whitespace().map(|n| n.parse().unwrap()).collect())
+                .collect()
+        };
+        let positions = vectors("v ");
+        let normals = vectors("vn ");
+        let face = obj.lines().find_map(|line| line.strip_prefix("f ")).unwrap();
+        let refs: Vec<Vec<usize>> = face.split_whitespace()
+            .map(|token| token.split("//").map(|n| n.parse::<usize>().unwrap() - 1).collect())
+            .collect();
+        assert_eq!(refs.len(), 3);
+        let a = &positions[refs[0][0]];
+        let b = &positions[refs[1][0]];
+        let c = &positions[refs[2][0]];
+        // Original face is +Z; the proper rotation takes its outward normal
+        // to +Y. Compute from exported vertex references, independently of
+        // the implementation's index ordering and normal conversion.
+        let cross_y = (b[2] - a[2]) * (c[0] - a[0])
+            - (b[0] - a[0]) * (c[2] - a[2]);
+        assert!(cross_y > 0.0, "exported face points inward: {face}");
+        for reference in refs {
+            if include_normals {
+                assert_eq!(reference.len(), 2);
+                assert_eq!(normals[reference[1]], [0.0, 1.0, 0.0]);
+            } else {
+                assert_eq!(reference.len(), 1);
+                assert!(normals.is_empty());
+            }
+        }
+    }
+}
+
+/// A minimal, otherwise-valid mesh — mirrors the fixture in
+/// `usd::tests::mesh_emittable_rejects_pathological_meshes`.
+fn good_mesh() -> MeshData {
+    MeshData {
+        express_id: 1,
+        ifc_type: "IfcWall".into(),
+        global_id: None,
+        name: None,
+        presentation_layer: None,
+        positions: vec![0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0, 0.0],
+        normals: vec![0.0, 0.0, 1.0, 0.0, 0.0, 1.0, 0.0, 0.0, 1.0],
+        indices: vec![0, 1, 2],
+        color: [0.5, 0.5, 0.5, 1.0],
+        material_name: None,
+        geometry_item_id: None,
+        material_id: None,
+        properties: None,
+        uvs: None,
+        texture: None,
+        geometry_class: 0,
+        origin: [0.0, 0.0, 0.0],
+        instance: None,
+        local_bounds: None,
+        local_to_world: None,
+    }
+}
+
+/// `mesh_visible` must reject a mesh carrying a non-finite position, normal, or
+/// origin — OBJ's `v`/`vn` tokens have no lexical form for `NaN`/`inf`/`-inf`, and
+/// unlike `mesh_input::scrub_nonfinite` (the from-meshes GLB/COLLADA gate) or
+/// `usd::mesh_emittable` (the sibling from-bytes exporter), this function let one
+/// through untouched. Before the fix each of these five cases wrote the offending
+/// float straight into a `v`/`vn` line (Rust's `Display` renders `NaN`/`inf`/`-inf`,
+/// none of which OBJ readers accept as a number).
+#[test]
+fn mesh_visible_rejects_non_finite_geometry() {
+    let good = good_mesh();
+    assert!(mesh_visible(&good, &None, &[]), "the baseline fixture must itself be visible");
+
+    let mut bad = good.clone();
+    bad.positions[0] = f32::NAN;
+    assert!(!mesh_visible(&bad, &None, &[]), "NaN position must be rejected");
+
+    let mut bad = good.clone();
+    bad.positions[3] = f32::INFINITY;
+    assert!(!mesh_visible(&bad, &None, &[]), "Infinity position must be rejected");
+
+    let mut bad = good.clone();
+    bad.normals[1] = f32::NEG_INFINITY;
+    assert!(!mesh_visible(&bad, &None, &[]), "non-finite normal must be rejected");
+
+    let mut bad = good.clone();
+    bad.origin = [f64::NAN, 0.0, 0.0];
+    assert!(!mesh_visible(&bad, &None, &[]), "non-finite origin must be rejected — it poisons every vertex");
+}
+
